Fix Virtio Ring Bug
The way in which descriptor items were presented to the device (and,
subsequently, retreived from the device) was incorrect: the avail ring
has a fixed number of items, but the value written into the field that
tells the device which descriptor items are available is essentially
unlimited. In other words, to set the ID of the descriptor item in the
avail run requires a mod by the ring size. However, that unmodified
value is what should be written into the avail index.

Signed-off-by: Will Hawkins <hawkinsw@obs.cr>